backend/requirements.txt
Django==6.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-spectacular==0.27.2
mysqlclient==2.2.8
python-dotenv==1.0.1
django-cors-headers==4.3.0